services/video-observation-mvp/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selfmeta-video-observation-mvp"
version = "0.1.0"
description = "Backend-first MVP for Self Meta free-play video observation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1",
  "pydantic>=2.11,<3",
  "sqlalchemy>=2.0,<3",
  "uvicorn>=0.35,<1",
  "pyyaml>=6.0,<7",
  "httpx>=0.28,<1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9"
]
media = [
  "mediapipe>=0.10.14,<0.11",
  "opencv-python-headless>=4.10,<5"
]
storage = [
  "boto3>=1.35,<2"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
